refactor(suite): use Collapsible in CollapsibleFees
