+++ packages/components/src/components/CollapsibleBox/CollapsibleBox.tsx
@@ -7,17 +7,22 @@ import {
     borders,
     mapElevationToBackground,
     mapElevationToBorder,
+    spacings,
     spacingsPx,
 } from '@trezor/theme';
 
-import { CollapsibleHeader } from './CollapsibleHeader';
+import { FillType, HeadingSize, PaddingType } from './types';
 import {
-    CollapsibleHeaderContent,
-    CollapsibleHeaderContentProps,
-} from './CollapsibleHeaderContent';
-import { FillType, PaddingType } from './types';
-import { mapPaddingTypeToContentPadding } from './utils';
+    mapPaddingTypeToContentPadding,
+    mapPaddingTypeToHeaderPadding,
+    mapSizeToHeadingTypography,
+    mapSizeToIconSize,
+    mapSizeToSubheadingTypography,
+} from './utils';
 import { Collapsible } from '../Collapsible/Collapsible';
+import { Column, Row } from '../Flex/Flex';
+import { IconName, IconProps, IconSize } from '../Icon/Icon';
+import { Text } from '../typography/Text/Text';
 import { ElevationUp, useElevation } from './../ElevationContext/ElevationContext';
 import {
     FrameProps,
@@ -39,6 +44,10 @@ type ContainerProps = {
     $fillType: FillType;
 };
 
+type HeaderProps = {
+    $paddingType: PaddingType;
+};
+
 type ContentProps = {
     $paddingType: PaddingType;
     $elevation: Elevation;
@@ -46,26 +55,21 @@ type ContentProps = {
 };
 
 export type CollapsibleBoxProps = AllowedFrameProps & {
+    heading: ReactNode;
+    subHeading?: ReactNode;
+    headingSize?: HeadingSize;
     paddingType?: PaddingType;
     fillType?: FillType;
+    toggleLabel?: ReactNode;
+    toggleIconName?: IconName;
+    toggleIconSize?: IconSize;
+    toggleIconVariant?: IconProps['variant'];
     children?: ReactNode;
     hasDivider?: boolean;
     onAnimationComplete?: (isOpen: boolean) => void;
     'data-testid'?: string;
-    'data-testid-toggle'?: string;
     defaultIsOpen?: boolean;
-    collapsible?: boolean;
-} & Pick<
-        CollapsibleHeaderContentProps,
-        | 'heading'
-        | 'subHeading'
-        | 'headingSize'
-        | 'toggleLabel'
-        | 'toggleComponent'
-        | 'toggleIconName'
-        | 'toggleIconSize'
-        | 'toggleIconVariant'
-    >;
+};
 
 const Container = styled.section<TransientProps<AllowedFrameProps> & ContainerProps>`
     width: 100%;
@@ -92,6 +96,21 @@ const Container = styled.section<TransientProps<AllowedFrameProps> & ContainerPr
     ${withFrameProps}
 `;
 
+const Toggle = styled.div`
+    transition: opacity 0.15s;
+`;
+
+const Header = styled.header<HeaderProps>`
+    padding: ${mapPaddingTypeToHeaderPadding};
+    cursor: pointer;
+
+    &:hover {
+        ${Toggle} {
+            opacity: 0.5;
+        }
+    }
+`;
+
 const Content = styled.div<ContentProps>`
     display: flex;
     flex-direction: column;
@@ -111,27 +130,65 @@ const Content = styled.div<ContentProps>`
 
 export const CollapsibleBox = ({
     defaultIsOpen = false,
+    toggleLabel,
+    toggleIconName = 'caretCircleDown',
     paddingType = 'normal',
     heading,
     subHeading,
-    headingSize,
-    toggleLabel,
-    toggleIconName,
-    toggleIconSize,
+    headingSize = 'large',
+    toggleIconSize = headingSize,
     toggleIconVariant,
-    toggleComponent,
     fillType = 'default',
     hasDivider = true,
     children,
     onAnimationComplete,
     'data-testid': dataTest,
-    'data-testid-toggle': dataTestToggleId = '@collapsible-box/toggle',
-    collapsible = true,
     ...rest
 }: CollapsibleBoxProps) => {
     const { elevation } = useElevation();
     const [isOpen, setIsOpen] = useState(defaultIsOpen);
     const frameProps = pickAndPrepareFrameProps(rest, allowedCollapsibleBoxFrameProps);
+
+    const headerContent = (
+        <Row gap={spacings.xs} justifyContent="space-between">
+            <Column alignItems="flex-start">
+                <Text
+                    as="div"
+                    typographyStyle={mapSizeToHeadingTypography({
+                        $headingSize: headingSize,
+                    })}
+                >
+                    {heading}
+                </Text>
+                {subHeading && (
+                    <Text
+                        as="div"
+                        typographyStyle={mapSizeToSubheadingTypography({
+                            $headingSize: headingSize,
+                        })}
+                        variant="tertiary"
+                    >
+                        {subHeading}
+                    </Text>
+                )}
+            </Column>
+            <Toggle>
+                <Row gap={spacings.sm}>
+                    {toggleLabel && (
+                        <Text typographyStyle="hint" variant="tertiary">
+                            {toggleLabel}
+                        </Text>
+                    )}
+                    <Collapsible.ToggleIcon
+                        iconName={toggleIconName}
+                        size={toggleIconSize ?? mapSizeToIconSize({ $headingSize: headingSize })}
+                        data-testid={`@collapsible-box/icon-${isOpen ? 'expanded' : 'collapsed'}`}
+                        variant={toggleIconVariant}
+                    />
+                </Row>
+            </Toggle>
+        </Row>
+    );
 
     return (
         <Container
@@ -141,32 +198,15 @@ export const CollapsibleBox = ({
             $fillType={fillType}
             data-testid={dataTest}
         >
-            <Collapsible isOpen={isOpen && collapsible}>
-                <Collapsible.Toggle
-                    data-testid={dataTestToggleId}
-                    onClick={() => {
-                        if (collapsible) {
-                            setIsOpen(!isOpen);
-                        }
-                    }}
-                >
-                    <CollapsibleHeader
-                        paddingType={paddingType}
-                        fillType={fillType}
-                        collapsible={collapsible}
-                    >
-                        <CollapsibleHeaderContent
-                            isOpen={isOpen}
-                            heading={heading}
-                            subHeading={subHeading}
-                            toggleLabel={toggleLabel}
-                            toggleComponent={toggleComponent}
-                            toggleIconName={toggleIconName}
-                            toggleIconSize={toggleIconSize}
-                            toggleIconVariant={toggleIconVariant}
-                            collapsible={collapsible}
-                        />
-                    </CollapsibleHeader>
+            <Collapsible isOpen={isOpen}>
+                <Collapsible.Toggle onClick={() => setIsOpen(!isOpen)}>
+                    <Header $paddingType={paddingType}>
+                        {fillType === 'none' ? (
+                            headerContent
+                        ) : (
+                            <ElevationUp>{headerContent}</ElevationUp>
+                        )}
+                    </Header>
                 </Collapsible.Toggle>
                 <Collapsible.Content
                     data-testid="@collapsible-box/body"
